src/requirements.txt
Django==2.2
djangorestframework==3.9.2
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.2
model-mommy==2.0.0
